feat: lengthen  cron duration and add command to trigger
