+++ apps/bot/src/lib/epic-helper/command-helper/utils/index.ts
@@ -1,0 +1,5 @@
+import {_syncUserReminderToDb} from './sync-user-reminder-to-db';
+
+export const _utilsHelper = {
+  syncReminderToDb: _syncUserReminderToDb,
+};
